Separate canvas react context from canvas components
