+++ assets/src/edit-story/app/canvas/context.js
