+++ assets/src/edit-story/app/canvas/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,5 @@
  * limitations under the License.
  */
 
-/**
- * Internal dependencies
- */
-import { identity, useContextSelector } from '../../../design-system';
-import Context from './context';
-
-function useCanvas(selector) {
-  return useContextSelector(Context, selector ?? identity);
-}
-
-export default useCanvas;
+export { default as CanvasProvider } from './canvasProvider';
+export { default as useCanvas } from './useCanvas';
