+++ assets/src/edit-story/app/canvas/utils/createPolygon.js
