+++ assets/src/edit-story/components/canvas/displayLayer.js
@@ -32,10 +32,9 @@ import {
   STORY_ANIMATION_STATE,
   useStoryAnimationContext,
 } from '../../../animation';
-import { useStory } from '../../app';
+import { useStory, useCanvas } from '../../app';
 import DisplayElement from './displayElement';
 import { Layer, PageArea } from './layout';
-import useCanvas from './useCanvas';
 import PageAttachment from './pageAttachment';
 
 function DisplayPage({
