+++ assets/src/edit-story/components/canvas/framesLayer.js
@@ -31,14 +31,13 @@ import { __ } from '@wordpress/i18n';
  */
 import { STORY_ANIMATION_STATE } from '../../../animation';
 import { PAGE_WIDTH, DESIGN_SPACE_MARGIN } from '../../constants';
-import { useStory, useDropTargets } from '../../app';
+import { useStory, useDropTargets, useCanvas } from '../../app';
 import withOverlay from '../overlay/withOverlay';
 import PageMenu from './pagemenu';
 import { Layer, MenuArea, NavNextArea, NavPrevArea, PageArea } from './layout';
 import FrameElement from './frameElement';
-import useCanvasKeys from './useCanvasKeys';
+import useCanvasKeys from '../../app/canvas/useCanvasKeys';
 import Selection from './selection';
-import useCanvas from './useCanvas';
 import PageNav from './pagenav';
 
 const FramesPageArea = withOverlay(
