+++ assets/src/edit-story/components/canvas/layout.js
@@ -39,7 +39,7 @@ import {
 } from '../../constants';
 import pointerEventsCss from '../../utils/pointerEventsCss';
 import generatePatternStyles from '../../utils/generatePatternStyles';
-import useCanvas from './useCanvas';
+import { useCanvas } from '../../app';
 
 /**
  * @file See https://user-images.githubusercontent.com/726049/72654503-bfffe780-3944-11ea-912c-fc54d68b6100.png
