+++ assets/src/edit-story/components/canvas/multiSelectionMoveable/index.js
@@ -24,13 +24,12 @@ import { useRef, useEffect, useState } from 'react';
  * Internal dependencies
  */
 import Moveable from '../../moveable';
-import { useStory } from '../../../app';
+import { useStory, useCanvas } from '../../../app';
 import objectWithout from '../../../utils/objectWithout';
 import { useTransform } from '../../transform';
 import { useUnits } from '../../../units';
 import { getDefinitionForType } from '../../../elements';
 import isTargetOutOfContainer from '../../../utils/isTargetOutOfContainer';
-import useCanvas from '../useCanvas';
 import useSnapping from '../utils/useSnapping';
 import useWindowResizeHandler from '../useWindowResizeHandler';
 import useDrag from './useDrag';
