+++ assets/src/edit-story/components/canvas/pageAttachment/test/pageAttachment.js
@@ -23,7 +23,7 @@ import React from 'react';
  * Internal dependencies
  */
 import PageAttachment from '../index';
-import CanvasContext from '../../context';
+import CanvasContext from '../../../../app/canvas/context';
 import { renderWithTheme } from '../../../../testUtils';
 import StoryContext from '../../../../app/story/context';
