+++ assets/src/edit-story/components/canvas/utils/useElementOutOfCanvas.js
@@ -17,8 +17,7 @@
  * Internal dependencies
  */
 import isTargetOutOfContainer from '../../../utils/isTargetOutOfContainer';
-import { useStory } from '../../../app/story';
-import useCanvas from '../useCanvas';
+import { useStory, useCanvas } from '../../../app/story';
 
 function useElementOutOfCanvas() {
   const { deleteSelectedElements } = useStory((state) => ({
