+++ assets/src/edit-story/components/panels/design/textStyle/test/textStyle.js
@@ -32,7 +32,7 @@ import DropDown from '../../../../form/dropDown';
 import AdvancedDropDown from '../../../../form/advancedDropDown';
 import ColorInput from '../../../../form/color/color';
 import createSolid from '../../../../../utils/createSolid';
-import CanvasContext from '../../../../canvas/context';
+import CanvasContext from '../../../../../app/canvas/context';
 import {
   HIDDEN_PADDING,
   MULTIPLE_VALUE,
